Fixed main file to run cli based on runtime argument
