+++ pom.xml
@@ -4,7 +4,7 @@
 
   <groupId>syrup</groupId>
   <artifactId>uml</artifactId>
-  <version>2</version>
+  <version>3</version>
   <packaging>jar</packaging>
 
   <name>uml-editor</name>
